Move hlsStrip and mp4Handler to helper
Also, disable all mp4 logic until SVT starts using the format again
(Homeland used this format for a while).
