+++ resources/lib/helper.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 import CommonFunctions
 import re
+import urllib
 
 common = CommonFunctions
 THUMB_SIZE = "extralarge"
@@ -42,7 +43,7 @@ def convertDuration(duration):
   1 min
   """
 
-  match = re.match(r'(^(\d+)\sh)*(\s*(\d+)\smin)*(\s*(\d+)\ssek)*',duration)
+  match = re.match(r'(^(\d+)\sh)*(\s*(\d+)\smin)*(\s*(\d+)\ssek)*', duration)
 
   dhours = 0
   dminutes = 0
@@ -81,26 +82,26 @@ def getUrlParameters(arguments):
   return params
 
 
-def tabExists(html,tabname):
+def tabExists(html, tabname):
   """
   Check if a specific tab exists in the DOM.
   """
-  return elementExists(html,"div",{ "data-tabname": tabname})
+  return elementExists(html, "div", { "data-tabname": tabname})
 
 
-def elementExists(html,etype,attrs):
+def elementExists(html, etype, attrs):
   """
   Check if a specific element exists in the DOM.
   """
 
-  htmlelement = common.parseDOM(html,etype, attrs = attrs)
+  htmlelement = common.parseDOM(html, etype, attrs = attrs)
 
   return len(htmlelement) > 0
 
 
 def prepareThumb(thumbnail):
   """
-  Returns a thumbnail with size 'large'
+  Returns a thumbnail with size THUMB_SIZE
   """
   common.log("old thumbnail: " + thumbnail)
   if not thumbnail.startswith("http://"):
@@ -108,3 +109,79 @@ def prepareThumb(thumbnail):
   thumbnail = re.sub(r"/small|medium|large|extralarge/", "/"+THUMB_SIZE+"/", thumbnail)
   common.log("new thumbnail: " + thumbnail)
   return thumbnail
+
+
+def mp4Handler(jsonObj):
+  """
+  Returns a mp4 stream URL.
+
+  If there are several mp4 streams in the JSON object:
+  pick the one with the highest bandwidth.
+
+  Some programs are available with multiple mp4 streams
+  for different bitrates. This function ensures that the one
+  with the highest bitrate is chosen.
+
+  Can possibly be extended to support some kind of quality
+  setting in the plugin.
+  """
+  videos = []
+
+  # Find all mp4 videos
+  for video in jsonObj["video"]["videoReferences"]:
+    if video["url"].endswith(".mp4"):
+      videos.append(video)
+  
+  if len(videos) == 1:
+    return videos[0]["url"]
+
+  bitrate = 0
+  url = ""
+
+  # Find the video with the highest bitrate
+  for video in videos:
+    if video["bitrate"] > bitrate:
+      bitrate = video["bitrate"]
+      url = video["url"]          
+
+  common.log("Info: bitrate="+str(bitrate)+" url="+url)
+  return url
+
+
+def hlsStrip(videoUrl):
+    """
+    Extracts the stream that supports the
+    highest bandwidth and is not using the avc1.77.30 codec.
+    """
+    common.log("Stripping file: " + videoUrl)
+
+    ufile = urllib.urlopen(videoUrl)
+    lines = ufile.readlines()
+
+    hlsurl = ""
+    bandwidth = 0
+    foundhigherquality = False
+
+    for line in lines:
+      if foundhigherquality:
+        # The stream url is on the line proceeding the header
+        foundhigherquality = False
+        hlsurl = line
+      if "EXT-X-STREAM-INF" in line: # The header
+        if not "avc1.77.30" in line:
+          match = re.match(r'.*BANDWIDTH=(\d+).+', line)
+          if match:
+            if bandwidth < int(match.group(1)):
+              foundhigherquality = True
+              bandwidth = int(match.group(1))
+          continue
+
+    if bandwidth == 0:
+      return None
+
+    ufile.close()
+    hlsurl = hlsurl.rstrip()
+    common.log("Returned stream url : " + hlsurl)
+    return hlsurl
+
+
